requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cssselect==1.3.0
eval-type-backport==0.2.2
faiss-cpu==1.10.0
frozenlist==1.5.0
goose3==3.1.19
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
langdetect==1.0.9
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
mistralai==1.7.0
multidict==6.4.3
numpy==2.2.4
packaging==24.2
pillow==11.2.1
propcache==0.3.1
pyahocorasick==2.1.0
pycares==4.6.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pythonmonkey-fork==1.1.0
pythonmonkey-node-modules==0.1.1
requests==2.32.3
rich==14.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
yarl==1.20.0
